Fix for PROD-3334, pull Volumes and Snapshots out of EC2Discovery and into its own
